+++ NEWS.md
@@ -57,6 +57,7 @@ __label__ have been removed
 
 
 ## Major changes
+- New `addF7Tooltip()` and `updateF7Tooltip()` to generate tooltips from the server
 - New __allowPWA__ parameter in `f7Page()` so that we doesn't oblige users to create a PWA
 - `f7Page()` has a new __options__ parameter as replacement of `f7Init()`. Much simpler to handle
 - update framework7 from 5.5.0 to 5.7.14